:recycle: move model -> ui/model directory
