+++ lib/application/pilgrimage/update_pilgrimage_progress_usecase.dart
@@ -4,7 +4,7 @@ import 'package:virtualpilgrimage/application/pilgrimage/update_pilgrimage_progr
 import 'package:virtualpilgrimage/application/pilgrimage/update_pilgrimage_progress_result.codegen.dart';
 import 'package:virtualpilgrimage/application/user/health/health_repository.dart';
 import 'package:virtualpilgrimage/application/user/user_repository.dart';
-import 'package:virtualpilgrimage/domain/pilgrimage/pilgrimage_progress_calculator.dart';
+import 'package:virtualpilgrimage/domain/pilgrimage/virtual_position_calculator.dart';
 import 'package:virtualpilgrimage/logger.dart';
 
 final updatePilgrimageProgressUsecaseProvider = Provider<UpdatePilgrimageProgressUsecase>(
