+++ lib/domain/pilgrimage/virtual_position_calculator.dart
