+++ test/application/pilgrimage/update_pilgrimage_progress_interactor_test.dart
@@ -11,8 +11,8 @@ import 'package:virtualpilgrimage/application/pilgrimage/update_pilgrimage_progr
 import 'package:virtualpilgrimage/application/user/health/health_repository.dart';
 import 'package:virtualpilgrimage/domain/customizable_date_time.dart';
 import 'package:virtualpilgrimage/domain/pilgrimage/pilgrimage_info.codegen.dart';
-import 'package:virtualpilgrimage/domain/pilgrimage/pilgrimage_progress_calculator.dart';
 import 'package:virtualpilgrimage/domain/pilgrimage/temple_info.codegen.dart';
+import 'package:virtualpilgrimage/domain/pilgrimage/virtual_position_calculator.dart';
 import 'package:virtualpilgrimage/domain/user/health/health_by_period.codegen.dart';
 import 'package:virtualpilgrimage/domain/user/virtual_pilgrimage_user.codegen.dart';
