+++ test/ui/model/form_model.codegen_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:virtualpilgrimage/model/form_model.codegen.dart';
+import 'package:virtualpilgrimage/ui/model/form_model.codegen.dart';
 
 void main() {
   group('FormModel', () {
